Refactor to feature first folder structure
