+++ src/components/layout/Layout.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { Outlet } from 'react-router-dom';
-import Navbar from 'components/Navbar';
+import Navbar from './Navbar';
 
 function Layout() {
   return (
