+++ src/components/layout/Navbar.module.css
@@ -1,0 +1,28 @@
+.header {
+  display: flex;
+  align-items: center;
+  gap: 2rem;
+  height: 4rem;
+  padding: 2rem;
+  background-color: #fff;
+  box-shadow: 1px 0 3px grey;
+}
+
+.logo {
+  font-size: 1.4rem;
+  font-weight: bold;
+  color: var(--color-azure);
+  text-decoration: none;
+}
+
+.navItems {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  list-style: none;
+  gap: 1rem;
+}
+
+.navLink {
+  text-decoration: none;
+}
